% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_helpers.R
\name{make_standard_folders}
\alias{make_standard_folders}
\title{Make folders for a project}
\usage{
make_standard_folders()
}
\value{
This function does not return anything. It sets up project folders
}
\description{
Make folders for a project
}
\examples{
make_standard_folders()
}
